requirements.txt
requests
beautifulsoup4
selenium==4.20.0
geckodriver-autoinstaller
simplejson
websocket-client